Fixed some stuff, I'm pretty sure
Fixed the turn text w.r.t. AI so that the turn text only changes when the camera starts to move (i.e. at the end, not beginning of the wait period before camera rotation).
Also, fixed it so that if the first player is an AI, they play first.